Abstract out validation logic
