+++ commands/records/create.js
@@ -1,6 +1,5 @@
 const righto = require('righto');
 const sqlite = require('sqlite-fp');
-const ErrorWithObject = require('error-with-object');
 const finalStream = require('final-stream');
 const writeResponse = require('write-response');
 
@@ -9,114 +8,10 @@ const getCollection = require('../../modules/getCollection');
 const getUser = require('../../modules/getUser');
 const applyPresentersToData = require('../../modules/applyPresentersToData');
 const applyTransducersToData = require('../../modules/applyTransducersToData');
-const evaluate = require('../../modules/evaluate');
 const writeResponseError = require('../../modules/writeResponseError');
+const validateDataAgainstSchema = require('../../modules/validateDataAgainstSchema');
 
 const uuidv4 = require('uuid').v4;
-
-function checkFieldValidation (schema, field, scope, callback) {
-  if (!schema[field]) {
-    return callback(null, { [field]: 'unknown column' });
-  }
-
-  const fieldValidations = schema[field].map(checkFn => {
-    if (checkFn === 'required') {
-      return;
-    }
-
-    if (checkFn === 'string') {
-      checkFn = 'value == null || getType(value) == "string" ? null : "must be string"';
-    }
-
-    if (checkFn === 'number') {
-      checkFn = 'value == null || getType(value) == "number" ? null : "must be number"';
-    }
-
-    if (checkFn === 'array') {
-      checkFn = 'value == null || getType(value) == "Array" ? null : "must be array"';
-    }
-
-    return righto(evaluate, checkFn, {
-      ...scope,
-      field,
-      value: scope.body[field]
-    });
-  });
-
-  righto.all(fieldValidations)(function (error, fieldResult) {
-    if (error) { return callback(error); }
-
-    const errors = fieldResult.reduce((accumulator, invalid) => {
-      if (invalid) {
-        accumulator[field] = accumulator[field] || [];
-        accumulator[field].push(invalid);
-      }
-      return accumulator;
-    }, {});
-
-    callback(null, errors);
-  });
-}
-
-function checkSchemaValidations (schema, scope, errors, callback) {
-  if (!schema) {
-    return callback();
-  }
-
-  const schemaValidations = Object.keys(scope.body)
-    .map(field => {
-      return righto(checkFieldValidation, schema, field, scope);
-    });
-
-  righto.all(schemaValidations)(function (error, results) {
-    if (error) { return callback(error); }
-
-    results.forEach(validationError => {
-      errors = Object.assign(errors, validationError);
-    });
-
-    if (Object.keys(errors).length > 0) {
-      return callback(new ErrorWithObject({
-        statusCode: 400,
-        friendly: errors
-      }));
-    }
-    callback();
-  });
-}
-
-function validateDataAgainstSchema (collectionConfig, scope, callback) {
-  const { schema } = collectionConfig;
-
-  if (!schema) {
-    return callback(null, scope.body);
-  }
-
-  const errors = {};
-
-  if (schema) {
-    // TODO: Abstract checkForRequiredFields()
-    Object.keys(schema).forEach(field => {
-      if (schema[field].includes('required') && !scope.body[field]) {
-        errors[field] = errors[field] || [];
-        errors[field].push('required');
-      }
-    });
-
-    // TODO: Abstract checkForUnknownFields()
-    Object.keys(scope.body).map(field => {
-      if (!schema[field]) {
-        errors[field] = errors[field] || [];
-        errors[field].push('unknown field');
-      }
-    });
-  }
-
-  const schemaValidated = righto(checkSchemaValidations, schema, scope, errors);
-  const result = righto.mate(scope.body, righto.after(schemaValidated));
-
-  result(callback);
-}
 
 function insertRecordIntoDatabase (collectionId, data, dbConnection, callback) {
   if (typeof data !== 'object') {
